src/requirements.txt
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
load-dotenv==0.1.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.4.0
tomli==2.0.1
tzdata==2024.1
Werkzeug==3.0.2
